starts the graphical user interface developed with shiny.

starts the graphical user interface developed with shiny.

sdmApp(
  maxRequestSize = 50,
  debug = FALSE,
  theme = "IHSN",
  ...,
  shiny.server = FALSE
)

Arguments

maxRequestSize
(numeric) number defining the maximum allowed filesize (in megabytes) for uploaded files, defaults to 50MB

debug
logical if TRUE, set shiny-debugging options

theme
select stylesheet for the interface.

...
arguments (e.g host) that are passed through runApp when starting the shiny application

shiny.server
Setting this parameter to TRUE will return the app in the form of an object rather than invoking it. This is useful for deploying sdcApp via shiny-server.

Value

starts the interactive graphical user interface which may be used to perform the anonymisation process.

Examples

if (FALSE) {
#load the package
library(sdmApp)
sdmApp()
}